login and register user added, editing and deleting just by the user who created comment, added user 1-many and advert to review

+++ reviews/views.py
@@ -1,16 +1,18 @@
 from rest_framework import permissions
+from rest_framework.exceptions import PermissionDenied
 from .models import Review
 from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from .serializers import ReviewSerializer
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated
+
 
 # Create your views here.
 
 
 class ReviewListView(APIView):
-  permission_classes = (IsAuthenticatedOrReadOnly,)
+  permission_classes = (IsAuthenticated,)
 
   def get(self,request):
     reviews = Review.objects.all()
@@ -18,6 +20,7 @@ class ReviewListView(APIView):
     return Response(serialized_reviews.data, status=status.HTTP_200_OK)
 
   def post(self,request):
+    request.data['owner'] = request.user.id
     review = ReviewSerializer(data=request.data)
     if review.is_valid():
       review.save()
@@ -27,7 +30,7 @@ class ReviewListView(APIView):
   
 
 class ReviewDetailView(APIView):
-  permission_classes = (IsAuthenticatedOrReadOnly,)
+  permission_classes = (IsAuthenticated,)
 
   def get(self,request,pk):
     review = Review.objects.get(id=pk)
@@ -37,6 +40,8 @@ class ReviewDetailView(APIView):
   def put(self,request,pk):
     review = Review.objects.get(id=pk)
     updated_review = ReviewSerializer(review,data= request.data)
+    if review.owner != request.user:
+      raise PermissionDenied()
     if updated_review.is_valid():
       updated_review.save()
       return Response(updated_review.data,status=status.HTTP_202_ACCEPTED)
@@ -46,7 +51,9 @@ class ReviewDetailView(APIView):
   def delete(self,request,pk):
     try:
       review = Review.objects.get(id=pk)
-      review.delete()
     except:
       return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+    if review.owner != request.user:
+      raise PermissionDenied()
+    review.delete()
     return Response(status=status.HTTP_204_NO_CONTENT)
